move `platform.persistent` package to `data.persistent`
