+++ app/shared/src/commonMain/kotlin/data/models/subject/SubjectManager.kt
@@ -28,6 +28,8 @@ import me.him188.ani.app.data.models.episode.EpisodeInfo
 import me.him188.ani.app.data.models.episode.episode
 import me.him188.ani.app.data.models.episode.isKnownOnAir
 import me.him188.ani.app.data.models.episode.type
+import me.him188.ani.app.data.persistent.asDataStoreSerializer
+import me.him188.ani.app.data.persistent.dataStores
 import me.him188.ani.app.data.repository.BangumiEpisodeRepository
 import me.him188.ani.app.data.repository.BangumiSubjectRepository
 import me.him188.ani.app.data.repository.setSubjectCollectionTypeOrDelete
@@ -39,8 +41,6 @@ import me.him188.ani.app.data.repository.toSubjectInfo
 import me.him188.ani.app.data.source.media.EpisodeCacheStatus
 import me.him188.ani.app.data.source.media.MediaCacheManager
 import me.him188.ani.app.platform.Context
-import me.him188.ani.app.platform.persistent.asDataStoreSerializer
-import me.him188.ani.app.platform.persistent.dataStores
 import me.him188.ani.app.session.SessionManager
 import me.him188.ani.app.tools.caching.ContentPolicy
 import me.him188.ani.app.tools.caching.LazyDataCache
